requirements.txt
streamlit
openai
requests
pandas
numpy
time
python-dotenv